requirements.txt
bs4
lxml
nltk
mysql-connector-python
flask
flask-session